// src/features/articleEditForm/model/consts/consts.ts
export enum ValidateArticleError {
    INCORRECT_TITLE = 'INCORRECT_TITLE',
    INCORRECT_SUBTITLE = 'INCORRECT_SUBTITLE',
    INCORRECT_TAGS = 'INCORRECT_TAGS',
    INCORRECT_BLOCKS = 'INCORRECT_BLOCKS',
    INCORRECT_FIRST_BLOCK = 'ONCORRECT_FIRST_BLOCK',
    INCORRECT_BLOCK = 'INCORRECT_BLOCK',
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR',
}